refactor: consolidate pa_mqa_logits_mxfp8 (schedule-free prefill/decode, merge tests)
